Remove library destructors. Remove library init/exit functions.
It is not sane to force closure on destruction: since GCC on Windows
does not support order for constructors and destructors, there is no way
to ensure that library A depending on library B gets destroyed before
library B, which can cause use-after-free situations, and random
crashes.

The client code should close its opened resources anyway.